# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/accommodations/": {
      "get": {
        "tags": [
          "accommodations"
        ],
        "summary": "Read Accommodations",
        "operationId": "read_accommodations_api_accommodations__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccommodationResponse"
                  },
                  "type": "array",
                  "title": "Response Read Accommodations Api Accommodations  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accommodations"
        ],
        "summary": "Create Accommodation",
        "operationId": "create_accommodation_api_accommodations__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccommodationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accommodations/{accommodation_id}": {
      "get": {
        "tags": [
          "accommodations"
        ],
        "summary": "Read Accommodation",
        "operationId": "read_accommodation_api_accommodations__accommodation_id__get",
        "parameters": [
          {
            "name": "accommodation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Accommodation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "accommodations"
        ],
        "summary": "Update Accommodation",
        "operationId": "update_accommodation_api_accommodations__accommodation_id__put",
        "parameters": [
          {
            "name": "accommodation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Accommodation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccommodationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "accommodations"
        ],
        "summary": "Delete Accommodation",
        "operationId": "delete_accommodation_api_accommodations__accommodation_id__delete",
        "parameters": [
          {
            "name": "accommodation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Accommodation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/accommodations/evict/{student_id}": {
      "delete": {
        "tags": [
          "accommodations"
        ],
        "summary": "Evict Student",
        "operationId": "evict_student_api_accommodations_evict__student_id__delete",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccommodationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/floors/": {
      "get": {
        "tags": [
          "floors"
        ],
        "summary": "Read Floors",
        "operationId": "read_floors_api_floors__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FloorResponse"
                  },
                  "type": "array",
                  "title": "Response Read Floors Api Floors  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "floors"
        ],
        "summary": "Create Floor",
        "operationId": "create_floor_api_floors__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/floors/{floor_id}": {
      "get": {
        "tags": [
          "floors"
        ],
        "summary": "Read Floor",
        "operationId": "read_floor_api_floors__floor_id__get",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Floor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "floors"
        ],
        "summary": "Update Floor",
        "operationId": "update_floor_api_floors__floor_id__put",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Floor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FloorCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "floors"
        ],
        "summary": "Delete Floor",
        "operationId": "delete_floor_api_floors__floor_id__delete",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Floor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/floors/by_dormitory/{dormitory_id}": {
      "get": {
        "tags": [
          "floors"
        ],
        "summary": "Read Floors By Dormitory",
        "operationId": "read_floors_by_dormitory_api_floors_by_dormitory__dormitory_id__get",
        "parameters": [
          {
            "name": "dormitory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dormitory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FloorResponse"
                  },
                  "title": "Response Read Floors By Dormitory Api Floors By Dormitory  Dormitory Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccommodationCreate": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "date_from": {
            "type": "string",
            "format": "date",
            "title": "Date From"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "room_id",
          "date_from"
        ],
        "title": "AccommodationCreate"
      },
      "AccommodationResponse": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "date_from": {
            "type": "string",
            "format": "date",
            "title": "Date From"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "room_id",
          "date_from"
        ],
        "title": "AccommodationResponse"
      },
...
      "FloorCreate": {
        "properties": {
          "dormitory_id": {
            "type": "integer",
            "title": "Dormitory Id"
          },
          "floor_number": {
            "type": "integer",
            "title": "Floor Number"
          }
        },
        "type": "object",
        "required": [
          "dormitory_id",
          "floor_number"
        ],
        "title": "FloorCreate"
      },
      "FloorResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "dormitory_id": {
            "type": "integer",
            "title": "Dormitory Id"
          },
          "floor_number": {
            "type": "integer",
            "title": "Floor Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dormitory_id",
          "floor_number"
        ],
        "title": "FloorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}